Change sysfs backend can deal with multi displays
The sysfs backends for get_brightness and set_brightness are able to deal with
multiple devices within the acpi backlight class.

get_brightness: only the value from the first device found in alphabetical order
                is returned.

set_brightness: redirects a printf of digit with the passed value to every
                device found in alphabetical order.

Also the sysfs method is now the default even if brightnessctl is installed as
brightnessctl has not cutted any release tag since feb 2020 and debian tends to
package only release tags, not from the master branch, the code kinda resembles
the internals of acpilight if it was written in shell instead of python.